refactor(react-ui-sfx): rename Hello→Morph; move Blob/Ghost out of experimental/
Sfx housekeeping bundled with this branch:
- `Hello` component renamed to `Morph`; `components/index.ts` updated.
- `experimental/Blob` and `experimental/Ghost` promoted to top-level
  `components/Blob` and `components/Ghost` directories (each with a local
  `index.ts`). Not added to the public namespace `components/index.ts`
  re-export — they remain story-only.
