+++ packages/ui/react-ui-sfx/src/components/Blob/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2025 DXOS.org
+//
+
+export * from './Blob';
