+++ packages/ui/react-ui-sfx/src/components/Ghost/Ghost.stories.tsx
@@ -9,7 +9,7 @@ import { DXOS } from '@dxos/brand';
 import { log } from '@dxos/log';
 import { withLayout, withTheme } from '@dxos/react-ui/testing';
 
-import { Ghost, type GhostController, type GhostProps } from './Ghost';
+import { Ghost, type GhostController, type GhostProps } from '../experimental./Ghost/Ghost';
 
 const DefaultStory = (props: Partial<GhostProps>) => {
   return (
