+++ packages/ui/react-ui-sfx/src/components/Ghost/Ghost.tsx
