+++ packages/ui/react-ui-sfx/src/components/index.ts
@@ -3,7 +3,7 @@
 //
 
 export * from './Chaos';
-export * from './Hello';
+export * from './Morph';
 export * from './Matrix';
 export * from './Oscilloscope';
 export * from './Pulse';
